rename classes and packages to plan gen, add test case package
