--- textdb/textdb-dataflow/src/main/java/edu/uci/ics/textdb/queryplanner/JsonPlannerTest.java
@@ -1,12 +1,0 @@
-package edu.uci.ics.textdb.queryplanner;
-
-
-public class JsonPlannerTest {
-    
-    public static void main(String[] args) throws Exception {
-        new JsonPlanner().generateQueryPlan(SampleJsonQuery.sampleJsonQueryRegexMatcher);
-        
-        
-    }
-
-}
