+++ textdb/textdb-dataflow/src/test/java/edu/uci/ics/textdb/dataflow/jsonplangen/SampleJsonQuery.java
@@ -1,4 +1,4 @@
-package edu.uci.ics.textdb.queryplanner;
+package edu.uci.ics.textdb.dataflow.jsonplangen;
 
 public class SampleJsonQuery {
